Don't generally require args and paths to be valid UTF-8, remove dependency on camino
OsStr and OsString is used instead.
"_utf8" versions of functions are replaced with their regular versions.
Paths are printed with {:?} in logs and errors.
NIXSA_BWRAP_ARGS is still required to be UTF-8.
quote() still returns String, but is expected to perform a proper conversion from OsString under the hood.

Closes #32.